+++ src/main/java/edu/kis/vh/nursery/stack/IntStackImplementation.java
@@ -1,4 +1,4 @@
-package edu.kis.vh.nursery.list;
+package edu.kis.vh.nursery.stack;
 
 public interface IntStackImplementation {
     /**
